split TransactionCost to its own module
