+++ runtime/src/lib.rs
@@ -78,6 +78,7 @@ pub mod status_cache;
 mod storable_accounts;
 pub mod tiered_storage;
 pub mod transaction_batch;
+pub mod transaction_cost;
 pub mod transaction_error_metrics;
 pub mod transaction_priority_details;
 mod verify_accounts_hash_in_background;
